requirements.txt
faiss_cpu>=1.11.0
gradio >= 5.31.0
joblib>=1.5.0
langchain>=0.3.25
langchain_ollama>=0.3.3
langchain_openai>=0.3.18
numpy>=2.2.6
pandas>=2.2.3
python-dotenv>=1.1.0
scikit_learn>=1.6.1
sentence_transformers>=4.1.0
spacy>=3.8.6
pt_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.8.0/pt_core_news_lg-3.8.0-py3-none-any.whl#sha256=2561c9a72a938d37141e9694e1a36d25061a44ce7e4f3bad2d3fa3bb836191af
tqdm>=4.67.1
transformers>=4.52.1
Unidecode>=1.4.0
